In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-
+#include <complex.h>
 /**
 * main - This is the main function where all the codes are written in
 * Return: zero(0)
@@ -7,23 +7,23 @@
 
 int main(void)
 {
-	int m, n, o, p;
+	long int m, n, o, p;
 
 	n = 1;
 	m = 2;
 
-	printf("%d, %d, ", n, m);
+	printf("%ld, %ld, ", n, m);
 	for (o = 2; o <= 48; o++)
 	{
 		p = n + m;
-		printf("%lld, ", p);
+		printf("%ld, ", p);
 
 		n = m;
 		m = p;
 
 		if (o == 48)
 		{
-			printf("%lld", p);
+			printf("%ld", p);
 		}
 	}
 	printf("\n");
